Word Change
Changed suitable to available in sentence. The sentence as is gives the impression that there IS an existing table at NCBI.  There will be cases where the table does not exist at NCBI but may exist at another database.  So we need to make it clear that NCBI is not assumed to be the only place a table might be.